generalise aligment methods and move to dut_analysis.py
